+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_application_security_groups_operations.py
@@ -24,7 +24,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
         response = self.client.application_security_groups.begin_delete(
             resource_group_name=resource_group.name,
             application_security_group_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -36,7 +36,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
         response = self.client.application_security_groups.get(
             resource_group_name=resource_group.name,
             application_security_group_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -58,7 +58,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
                 "tags": {"str": "str"},
                 "type": "str",
             },
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -71,7 +71,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
             resource_group_name=resource_group.name,
             application_security_group_name="str",
             parameters={"tags": {"str": "str"}},
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself
@@ -81,7 +81,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
     @recorded_by_proxy
     def test_application_security_groups_list_all(self, resource_group):
         response = self.client.application_security_groups.list_all(
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself
@@ -92,7 +92,7 @@ class TestNetworkManagementApplicationSecurityGroupsOperations(AzureMgmtRecorded
     def test_application_security_groups_list(self, resource_group):
         response = self.client.application_security_groups.list(
             resource_group_name=resource_group.name,
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself