+++ sdk/resources/azure-mgmt-resource/generated_tests/test_deployment_stacks_deployment_stacks_operations.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 import pytest
-from azure.mgmt.resource import DeploymentStacksClient
+from azure.mgmt.resource.deploymentstacks.v2022_08_01_preview import DeploymentStacksClient
 
 from devtools_testutils import AzureMgmtRecordedTestCase, RandomNameResourceGroupPreparer, recorded_by_proxy