+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/_data_lake_file_client.py
@@ -461,7 +461,7 @@ class DataLakeFileClient(PathClient):
             A callback to track the progress of a long-running upload. The signature is
             function(current: int, total: int) where current is the number of bytes transferred
             so far, and total is the total size of the download.
-        :paramtype progress_hook: Callable[[int, int], None]
+        :paramtype progress_hook: ~typing.Callable[[int, int], None]
         :returns: A dictionary of response headers.
         :rtype: Dict[str, Any]
         """
@@ -692,7 +692,7 @@ class DataLakeFileClient(PathClient):
             A callback to track the progress of a long-running download. The signature is
             function(current: int, total: int) where current is the number of bytes transferred
             so far, and total is the total size of the download.
-        :paramtype progress_hook: Callable[[int, int], None]
+        :paramtype progress_hook: ~typing.Callable[[int, int], None]
         :keyword int timeout:
             Sets the server-side timeout for the operation in seconds. For more details see
             https://learn.microsoft.com/rest/api/storageservices/setting-timeouts-for-blob-service-operations.
@@ -821,8 +821,9 @@ class DataLakeFileClient(PathClient):
         :param str query_expression:
             Required. a query statement.
             eg. Select * from DataLakeStorage
-        :keyword Callable[~azure.storage.filedatalake.DataLakeFileQueryError] on_error:
+        :keyword on_error:
             A function to be called on any processing errors returned by the service.
+        :paramtype on_error: ~typing.Callable[~azure.storage.filedatalake.DataLakeFileQueryError]
         :keyword file_format:
             Optional. Defines the serialization of the data currently stored in the file. The default is to
             treat the file data as CSV data formatted in the default dialect. This can be overridden with