+++ sdk/healthdataaiservices/azure-health-deidentification/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0 (Unreleased)
+## 1.0.0 (2025-05-19)
 
 ### Features Added